requirements.txt
aiofiles==23.2.1
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.1
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
numpy==1.24.4
orjson==3.9.8
packaging==23.2
prompt-toolkit==3.0.39
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
sox==1.4.1
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
confluent-kafka==2.2.0
